rename `lp-list` to `right-bar/lp-list`
