+++ packages/plugins/io-center/src/query.svelte.ts
@@ -12,7 +12,7 @@ import {
 import { store } from "./store.svelte";
 import type { Nullable } from "./types";
 import type { Connection, ConnectionPort, LogicalConditioner } from "./ui/components/canvas/types.canvas";
-import type { LpElement } from "./ui/components/lp-list/types.lp-list";
+import type { LpElement } from "./ui/components/right-bar/lp-list/types.lp-list";
 
 
 /**
